pull some more wind functions into lib
